+++ src/Language/Alloy/Debug.hs
@@ -1,0 +1,14 @@
+{-|
+Module      : Language.Alloy.Call
+Copyright   : (c) Marcellus Siegburg, 2019 - 2021
+License     : MIT
+
+This module provides functions to retrieve raw instances form Alloy,
+i.e. as Alloy provides them.
+-}
+module Language.Alloy.Debug (
+  getRawInstances,
+  getRawInstancesWith,
+  ) where
+
+import Language.Alloy.Internal.Call     (getRawInstances, getRawInstancesWith)